Wendy Version 2.1 =================
* fixed (known) issues concerning reduction rule "succeeding sending event"
  (parameter: `--succeedingSendingEvent')

* issues on the hidden parameter "quit as soon as possible" (parameter:
  `--succeedingSendingEvent') have been fixed

* implemented on-the-fly knowledge analysis

* added test case with insane nontrivial TSCC

* integrated LoLA SCCs detection within inner markings

* generated temporary files are always deleted unless the `--noClean'
  parameter is given; removed parameter `--clean'

* fixed 64-bit support: Wendy natively supports 32 and 64 bit compilations
  using GCC's switches `-m32' and `-m64', respectively. To successfully
  execute `make check', you need to compile the Petri Net API and Mia with
  the same options.

* allocated memory can be released using the (undocumented) parameter
  `--finalize'; fixed most memory leaks

* simplified interface to `Output' class managing file output

+++ wendy/src/Cover.cc
@@ -23,6 +23,7 @@
 #include "InnerMarking.h"
 #include "Label.h"
 #include "verbose.h"
+#include "types.h"
 
 
 extern gengetopt_args_info args_info;
@@ -487,10 +488,10 @@ void Cover::write(std::ostream &os)
         os << "    (";
         
         if(current[j].size() > 0)
-          os << reinterpret_cast<unsigned long>(current[j][0]);
+          os << reinterpret_cast<size_t>(current[j][0]);
         for(unsigned int k=1; k<current[j].size(); ++k)
         {
-          os << " + " << reinterpret_cast<unsigned long>(current[j][k]); 
+          os << " + " << reinterpret_cast<size_t>(current[j][k]); 
         }
         
         if(--lastClause > 0)
@@ -519,7 +520,7 @@ void Cover::write(std::ostream &os)
     else
       firstBubble = false;
     
-    os << "  " << reinterpret_cast<unsigned long>(it->first) << " :";
+    os << "  " << reinterpret_cast<size_t>(it->first) << " :";
     if( it->second.inP.empty() && 
         it->second.comP.empty() &&
         it->second.inT.empty() &&
